Reintroduced full IDeploymentRepository interface after splitting it into config and application repositories
