+++ test/NugetFeedDeploymentRepositoryTest/NugetFeedDeploymentRepositoryTest.csproj
@@ -57,7 +57,7 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="NugetFeedDeploymentRepositoryTest.cs" />
+    <Compile Include="NugetFeedApplicationRepositoryTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
